msm8998-common:  overlay: force the 'natural' color mode Natural color mode: persist.sys.sf.color_saturation=1.0 Boosted color mode: persist.sys.sf.color_saturation=1.1
It may happen (for some reason) that the property
persist.sys.sf.color_saturation is set to 1.1. In this case, DC Dimming
causes the screen to go full black and white.

To avoid this issue, force the display color to always be 'normal',
allowing it to be the only valid color mode.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>